separate file for each mapper
