+++ src/utils/mappers/image-mapper.ts
@@ -1,0 +1,10 @@
+import { ImageType } from "@/domain/components/image-type";
+import { ImageResponse } from "@/domain/response/cms/image-response";
+
+export const mapImageResponse = (imageResponse: ImageResponse): ImageType => {
+  const { sys } = imageResponse;
+  return {
+    id: sys.id,
+    ...imageResponse,
+  };
+};
